requirements.txt
langchain 
wikipedia 
langchain-community 
langchain-anthropic
python-dotenv
pydantic
duckduckgo-search
Flask
Flask-Cors
anthropic
langchain-openai
boto3
flask